PLIC Driver Bug Fix and nk_dev_info_read_register_block Type Change
Found an off-by-one error in the PLIC driver caused by the fact PLIC
IRQ 0 is invalid.

Changed the "size" of register blocks returned from nk_dev_info
structures to be a size_t instead of an int.

+++ src/dev/gicv2.c
@@ -321,7 +321,7 @@ static int gicv2_init_dev_info(struct nk_dev_info *info)
   global_gicv2_ptr = gic;
 
   void * bases[2];
-  int sizes[2];
+  size_t sizes[2];
 
   if(nk_dev_info_read_register_blocks_exact(info, bases, sizes, 2)) {
     GIC_ERROR("Failed to read register mmio bases!\n");
